refactor(syscalls): better syscall module structure
